Move VmState to separate file
